remove warning from LLM preference & add warning to vector database selection
