+++ frontend/src/pages/GeneralSettings/VectorDatabase/index.jsx
@@ -12,6 +12,7 @@ import WeaviateLogo from "../../../media/vectordbs/weaviate.png";
 import QDrantLogo from "../../../media/vectordbs/qdrant.png";
 import PreLoader from "../../../components/Preloader";
 import VectorDBOption from "../../../components/VectorDBOption";
+import { Warning } from "@phosphor-icons/react";
 
 export default function GeneralVectorDatabase() {
   const [saving, setSaving] = useState(false);
@@ -35,24 +36,57 @@ export default function GeneralVectorDatabase() {
     setVectorDB(selection);
   };
 
+  // const handleSubmit = async (e) => {
+  //   e.preventDefault();
+  //   setSaving(true);
+  //   const data = {};
+  //   const form = new FormData(e.target);
+  //   for (var [key, value] of form.entries()) data[key] = value;
+  //   const { error } = await System.updateSystem(data);
+  //   if (error) {
+  //     showToast(`Failed to save settings: ${error}`, "error");
+  //   } else {
+  //     showToast("Settings saved successfully.", "success");
+  //   }
+  //   setSaving(false);
+  //   setHasChanges(!!error);
+  // };
+
   const handleSubmit = async (e) => {
     e.preventDefault();
-    setSaving(true);
-    const data = {};
-    const form = new FormData(e.target);
-    for (var [key, value] of form.entries()) data[key] = value;
-    const { error } = await System.updateSystem(data);
-    if (error) {
-      showToast(`Failed to save settings: ${error}`, "error");
+    if (vectorDB !== settings?.VectorDB && hasChanges) {
+      document.getElementById("confirmation-modal")?.showModal();
     } else {
-      showToast("Settings saved successfully.", "success");
+      await handleSaveSettings();
+    }
+  };
+
+  const handleSaveSettings = async () => {
+    setSaving(true);
+    const data = new FormData(document.getElementById("vectordb-form"));
+    const settingsData = {};
+    for (let [key, value] of data.entries()) {
+      settingsData[key] = value;
+    }
+
+    const { error } = await System.updateSystem(settingsData);
+    if (error) {
+      showToast(`Failed to save LLM settings: ${error}`, "error");
+      setHasChanges(true);
+    } else {
+      showToast("LLM preferences saved successfully.", "success");
+      setHasChanges(false);
     }
     setSaving(false);
-    setHasChanges(!!error);
+    document.getElementById("confirmation-modal")?.close();
   };
 
   return (
     <div className="w-screen h-screen overflow-hidden bg-sidebar flex">
+            <ConfirmationModal
+        onClose={() => document.getElementById("confirmation-modal")?.close()}
+        onConfirm={handleSaveSettings}
+      />
       {!isMobile && <Sidebar />}
       {loading ? (
         <div
@@ -70,6 +104,7 @@ export default function GeneralVectorDatabase() {
         >
           {isMobile && <SidebarMobileHeader />}
           <form
+            id="vectordb-form"
             onSubmit={handleSubmit}
             onChange={() => setHasChanges(true)}
             className="flex w-full"
@@ -337,3 +372,46 @@ export default function GeneralVectorDatabase() {
     </div>
   );
 }
+
+const ConfirmationModal = ({ onClose, onConfirm }) => (
+  <dialog id="confirmation-modal" className="bg-transparent outline-none">
+    <div className="relative w-full max-w-2xl max-h-full">
+      <div className="relative bg-main-gradient rounded-lg shadow">
+        <div className="flex items-start justify-between p-4 border-b rounded-t border-gray-500/50">
+          <div className="flex items-center gap-2">
+            <Warning
+              className="text-yellow-300 text-lg w-6 h-6"
+              weight="fill"
+            />
+            <h3 className="text-xl font-semibold text-yellow-300">Warning</h3>
+          </div>
+        </div>
+        <div className="w-[550px] p-6 text-white">
+          <p>
+            Switching the vector database may affect querying documents and similarity
+            search results.
+            <br />
+            <br />
+            Are you sure you want to proceed?
+          </p>
+        </div>
+
+        <div className="flex w-full justify-between items-center p-6 space-x-2 border-t rounded-b border-gray-500/50">
+          <button
+            onClick={onClose}
+            type="button"
+            className="px-4 py-2 rounded-lg text-white hover:bg-red-500 transition-all duration-300"
+          >
+            Cancel
+          </button>
+          <button
+            onClick={onConfirm}
+            className="transition-all duration-300 border border-slate-200 px-4 py-2 rounded-lg text-white text-sm items-center flex gap-x-2 hover:bg-slate-200 hover:text-slate-800 focus:ring-gray-800"
+          >
+            Confirm
+          </button>
+        </div>
+      </div>
+    </div>
+  </dialog>
+);
